# apis/backend-api/communication/sdkwork-customerservice-backend-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Customer Service Backend API",
    "version": "1.0.0",
    "description": "Operator-facing customer service ticket administration API.",
    "x-sdkwork-api-authority": "sdkwork-customerservice-backend-api",
    "x-sdkwork-sdk-family": "sdkwork-customerservice-backend-sdk",
    "x-sdkwork-owner": "sdkwork-customerservice",
    "x-sdkwork-domain": "communication",
    "x-sdkwork-capability": "customerservice"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18091"
    }
  ],
...
  "security": [
    {
      "AuthToken": [],
      "AccessToken": []
    }
  ],
  "paths": {
    "/backend/v3/api/customer_services/tickets": {
      "get": {
        "tags": [
          "customerServiceTicketsAdmin"
        ],
        "operationId": "customerservice.tickets.admin.list",
        "x-sdkwork-owner": "sdkwork-customerservice",
        "x-sdkwork-api-authority": "sdkwork-customerservice-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-customerservice-backend-api",
        "x-sdkwork-permission": "customerservice.tickets.admin.read",
        "x-sdkwork-auth-mode": "dual-token",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "additionalProperties": false,
                          "required": [
                            "items",
                            "pageInfo"
                          ],
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/TicketSummary"
                              }
                            },
                            "pageInfo": {
                              "type": "object",
                              "additionalProperties": false,
                              "required": [
                                "mode",
                                "hasMore"
                              ],
                              "properties": {
                                "mode": {
                                  "type": "string",
                                  "enum": [
                                    "cursor"
                                  ]
                                },
                                "nextCursor": {
                                  "type": [
                                    "string",
                                    "null"
                                  ]
                                },
                                "hasMore": {
                                  "type": "boolean"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ]
      }
    },
    "/backend/v3/api/customer_services/tickets/{ticketId}": {
      "get": {
        "tags": [
          "customerServiceTicketsAdmin"
        ],
        "operationId": "customerservice.tickets.admin.retrieve",
        "x-sdkwork-owner": "sdkwork-customerservice",
        "x-sdkwork-api-authority": "sdkwork-customerservice-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-customerservice-backend-api",
        "x-sdkwork-permission": "customerservice.tickets.admin.read",
        "x-sdkwork-auth-mode": "dual-token",
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/TicketDetail"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ]
      },
      "patch": {
        "tags": [
          "customerServiceTicketsAdmin"
        ],
        "operationId": "customerservice.tickets.admin.update",
        "x-sdkwork-owner": "sdkwork-customerservice",
        "x-sdkwork-api-authority": "sdkwork-customerservice-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-customerservice-backend-api",
        "x-sdkwork-permission": "customerservice.tickets.admin.write",
        "x-sdkwork-auth-mode": "dual-token",
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/TicketDetail"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ]
      }
    },
...
    "/backend/v3/api/customer_services/plugins": {
      "get": {
        "tags": [
          "customerServicePluginsAdmin"
        ],
        "operationId": "customerservice.plugins.admin.list",
        "x-sdkwork-owner": "sdkwork-customerservice",
        "x-sdkwork-api-authority": "sdkwork-customerservice-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-customerservice-backend-api",
        "x-sdkwork-permission": "customerservice.plugins.admin.read",
        "x-sdkwork-auth-mode": "dual-token",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "additionalProperties": false,
                          "required": [
                            "items",
                            "pageInfo"
                          ],
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/PluginCatalogEntry"
                              }
                            },
                            "pageInfo": {
                              "$ref": "#/components/schemas/PageInfo"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ]
      }
    },
...
    "/backend/v3/api/customer_services/channels/accounts/{accountId}/credentials": {
      "post": {
        "tags": [
          "customerServiceChannelsAdmin"
        ],
        "operationId": "customerservice.channels.admin.accounts.credentials.register",
        "x-sdkwork-owner": "sdkwork-customerservice",
        "x-sdkwork-api-authority": "sdkwork-customerservice-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-customerservice-backend-api",
        "x-sdkwork-permission": "customerservice.channels.admin.write",
        "x-sdkwork-auth-mode": "dual-token",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterChannelCredentialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkCommandResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    },
    "schemas": {
...
      "UpdateTicketRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "assigneeUserId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
    }
  }
}